migrate sign api to v2
