+++ httpapi/coreapi/accounts_api_test.go
@@ -12,7 +12,6 @@ import (
 	"testing"
 	"time"
 
-	coredocumentpb "github.com/centrifuge/centrifuge-protobufs/gen/go/coredocument"
 	"github.com/centrifuge/go-centrifuge/config"
 	"github.com/centrifuge/go-centrifuge/config/configstore"
 	"github.com/centrifuge/go-centrifuge/errors"
@@ -30,74 +29,6 @@ const (
 	p2pPub     = "../../build/resources/p2pKey.pub.pem"
 )
 
-func TestHandler_SignPayload(t *testing.T) {
-	getHTTPReqAndResp := func(ctx context.Context, b io.Reader) (*httptest.ResponseRecorder, *http.Request) {
-		return httptest.NewRecorder(), httptest.NewRequest("POST", "/accounts/{account_id}/sign", b).WithContext(ctx)
-	}
-	// empty account_id
-	rctx := chi.NewRouteContext()
-	rctx.URLParams.Keys = make([]string, 1, 1)
-	rctx.URLParams.Values = make([]string, 1, 1)
-	rctx.URLParams.Keys[0] = accountIDParam
-	rctx.URLParams.Values[0] = ""
-	ctx := context.WithValue(context.Background(), chi.RouteCtxKey, rctx)
-	w, r := getHTTPReqAndResp(ctx, nil)
-	h := handler{}
-	h.SignPayload(w, r)
-	assert.Equal(t, w.Code, http.StatusBadRequest)
-	assert.Contains(t, w.Body.String(), ErrAccountIDInvalid.Error())
-
-	// invalid account id
-	rctx.URLParams.Values[0] = "invalid value"
-	w, r = getHTTPReqAndResp(ctx, nil)
-	h.SignPayload(w, r)
-	assert.Equal(t, w.Code, http.StatusBadRequest)
-	assert.Contains(t, w.Body.String(), ErrAccountIDInvalid.Error())
-
-	// empty body
-	accountID := utils.RandomSlice(20)
-	rctx.URLParams.Values[0] = hexutil.Encode(accountID)
-	w, r = getHTTPReqAndResp(ctx, nil)
-	h.SignPayload(w, r)
-	assert.Equal(t, w.Code, http.StatusBadRequest)
-	assert.Contains(t, w.Body.String(), "unexpected end of JSON input")
-
-	// failed signer
-	payload := utils.RandomSlice(32)
-	body := map[string]string{
-		"payload": hexutil.Encode(payload),
-	}
-	d, err := json.Marshal(body)
-	assert.NoError(t, err)
-	srv := new(configstore.MockService)
-	srv.On("Sign", accountID, payload).Return(nil, errors.New("failed to sign payload")).Once()
-	h.srv.accountsSrv = srv
-	w, r = getHTTPReqAndResp(ctx, bytes.NewReader(d))
-	h.SignPayload(w, r)
-	assert.Equal(t, w.Code, http.StatusBadRequest)
-	assert.Contains(t, w.Body.String(), "failed to sign payload")
-	srv.AssertExpectations(t)
-
-	// success
-	signature := utils.RandomSlice(32)
-	pk := utils.RandomSlice(20)
-	srv = new(configstore.MockService)
-	srv.On("Sign", accountID, payload).Return(&coredocumentpb.Signature{
-		SignerId:  accountID,
-		Signature: signature,
-		PublicKey: pk,
-	}, nil).Once()
-	h.srv.accountsSrv = srv
-	w, r = getHTTPReqAndResp(ctx, bytes.NewReader(d))
-	h.SignPayload(w, r)
-	assert.Equal(t, w.Code, http.StatusOK)
-	assert.Contains(t, w.Body.String(), hexutil.Encode(payload))
-	assert.Contains(t, w.Body.String(), hexutil.Encode(signature))
-	assert.Contains(t, w.Body.String(), hexutil.Encode(pk))
-	assert.Contains(t, w.Body.String(), hexutil.Encode(accountID))
-	srv.AssertExpectations(t)
-}
-
 func TestHandler_GetAccount(t *testing.T) {
 	getHTTPReqAndResp := func(ctx context.Context) (*httptest.ResponseRecorder, *http.Request) {
 		return httptest.NewRecorder(), httptest.NewRequest("GET", "/accounts/{account_id}", nil).WithContext(ctx)
@@ -106,7 +37,7 @@ func TestHandler_GetAccount(t *testing.T) {
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Keys = make([]string, 1, 1)
 	rctx.URLParams.Values = make([]string, 1, 1)
-	rctx.URLParams.Keys[0] = accountIDParam
+	rctx.URLParams.Keys[0] = AccountIDParam
 	rctx.URLParams.Values[0] = ""
 	ctx := context.WithValue(context.Background(), chi.RouteCtxKey, rctx)
 	w, r := getHTTPReqAndResp(ctx)
@@ -410,7 +341,7 @@ func TestHandler_UpdateAccount(t *testing.T) {
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Keys = make([]string, 1, 1)
 	rctx.URLParams.Values = make([]string, 1, 1)
-	rctx.URLParams.Keys[0] = accountIDParam
+	rctx.URLParams.Keys[0] = AccountIDParam
 	rctx.URLParams.Values[0] = ""
 	ctx := context.WithValue(context.Background(), chi.RouteCtxKey, rctx)
 	h := handler{}
