+++ httpapi/v2/accounts_api_test.go
@@ -11,8 +11,10 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	coredocumentpb "github.com/centrifuge/centrifuge-protobufs/gen/go/coredocument"
 	"github.com/centrifuge/go-centrifuge/config/configstore"
 	"github.com/centrifuge/go-centrifuge/errors"
+	"github.com/centrifuge/go-centrifuge/httpapi/coreapi"
 	"github.com/centrifuge/go-centrifuge/utils"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/go-chi/chi"
@@ -64,4 +66,72 @@ func TestHandler_GenerateAccount(t *testing.T) {
 	assert.Equal(t, w.Code, http.StatusCreated)
 	assert.Contains(t, w.Body.String(), hexutil.Encode(did))
 	assert.Contains(t, w.Body.String(), hexutil.Encode(jobID))
+}
+
+func TestHandler_SignPayload(t *testing.T) {
+	getHTTPReqAndResp := func(ctx context.Context, b io.Reader) (*httptest.ResponseRecorder, *http.Request) {
+		return httptest.NewRecorder(), httptest.NewRequest("POST", "/accounts/{account_id}/sign", b).WithContext(ctx)
+	}
+	// empty account_id
+	rctx := chi.NewRouteContext()
+	rctx.URLParams.Keys = make([]string, 1)
+	rctx.URLParams.Values = make([]string, 1)
+	rctx.URLParams.Keys[0] = coreapi.AccountIDParam
+	rctx.URLParams.Values[0] = ""
+	ctx := context.WithValue(context.Background(), chi.RouteCtxKey, rctx)
+	w, r := getHTTPReqAndResp(ctx, nil)
+	h := handler{}
+	h.SignPayload(w, r)
+	assert.Equal(t, w.Code, http.StatusBadRequest)
+	assert.Contains(t, w.Body.String(), coreapi.ErrAccountIDInvalid.Error())
+
+	// invalid account id
+	rctx.URLParams.Values[0] = "invalid value"
+	w, r = getHTTPReqAndResp(ctx, nil)
+	h.SignPayload(w, r)
+	assert.Equal(t, w.Code, http.StatusBadRequest)
+	assert.Contains(t, w.Body.String(), coreapi.ErrAccountIDInvalid.Error())
+
+	// empty body
+	accountID := utils.RandomSlice(20)
+	rctx.URLParams.Values[0] = hexutil.Encode(accountID)
+	w, r = getHTTPReqAndResp(ctx, nil)
+	h.SignPayload(w, r)
+	assert.Equal(t, w.Code, http.StatusBadRequest)
+	assert.Contains(t, w.Body.String(), "unexpected end of JSON input")
+
+	// failed signer
+	payload := utils.RandomSlice(32)
+	body := map[string]string{
+		"payload": hexutil.Encode(payload),
+	}
+	d, err := json.Marshal(body)
+	assert.NoError(t, err)
+	srv := new(configstore.MockService)
+	srv.On("Sign", accountID, payload).Return(nil, errors.New("failed to sign payload")).Once()
+	h.srv.accountSrv = srv
+	w, r = getHTTPReqAndResp(ctx, bytes.NewReader(d))
+	h.SignPayload(w, r)
+	assert.Equal(t, w.Code, http.StatusBadRequest)
+	assert.Contains(t, w.Body.String(), "failed to sign payload")
+	srv.AssertExpectations(t)
+
+	// success
+	signature := utils.RandomSlice(32)
+	pk := utils.RandomSlice(20)
+	srv = new(configstore.MockService)
+	srv.On("Sign", accountID, payload).Return(&coredocumentpb.Signature{
+		SignerId:  accountID,
+		Signature: signature,
+		PublicKey: pk,
+	}, nil).Once()
+	h.srv.accountSrv = srv
+	w, r = getHTTPReqAndResp(ctx, bytes.NewReader(d))
+	h.SignPayload(w, r)
+	assert.Equal(t, w.Code, http.StatusOK)
+	assert.Contains(t, w.Body.String(), hexutil.Encode(payload))
+	assert.Contains(t, w.Body.String(), hexutil.Encode(signature))
+	assert.Contains(t, w.Body.String(), hexutil.Encode(pk))
+	assert.Contains(t, w.Body.String(), hexutil.Encode(accountID))
+	srv.AssertExpectations(t)
 }
