+++ httpapi/v2/handler.go
@@ -38,4 +38,5 @@ func Register(ctx map[string]interface{}, r chi.Router) {
 	r.Delete("/documents/{"+coreapi.DocumentIDParam+"}/attributes/{"+AttributeKeyParam+"}", h.DeleteAttribute)
 	r.Post("/accounts/generate", h.GenerateAccount)
 	r.Get("/jobs/{"+jobIDParam+"}", h.Job)
+	r.Post("/accounts/{"+coreapi.AccountIDParam+"}/sign", h.SignPayload)
 }
