+++ httpapi/v2/handler_test.go
@@ -13,5 +13,5 @@ func TestRegister(t *testing.T) {
 	r := chi.NewRouter()
 	ctx := map[string]interface{}{BootstrappedService: Service{}}
 	Register(ctx, r)
-	assert.Len(t, r.Routes(), 18)
+	assert.Len(t, r.Routes(), 19)
 }
